refactor(editor): Extract sourceControl code into features
