+++ packages/frontend/editor-ui/src/features/sourceControl.ee/components/SourceControlPullModal.test.ts
@@ -1,9 +1,9 @@
-import SourceControlPullModalEe from './SourceControlPullModal.ee.vue';
+import SourceControlPullModalEe from './SourceControlPullModal.vue';
 import { createComponentRenderer } from '@/__tests__/render';
 import { createTestingPinia } from '@pinia/testing';
 import { createEventBus } from '@n8n/utils/event-bus';
 import userEvent from '@testing-library/user-event';
-import { useSourceControlStore } from '@/stores/sourceControl.store';
+import { useSourceControlStore } from '../sourceControl.store';
 import { mockedStore } from '@/__tests__/utils';
 import { waitFor } from '@testing-library/dom';
 import { reactive } from 'vue';
