+++ packages/frontend/editor-ui/src/features/sourceControl.ee/components/SourceControlPushModal.test.ts
@@ -1,11 +1,11 @@
 import { within, waitFor } from '@testing-library/dom';
 import userEvent from '@testing-library/user-event';
 import { createComponentRenderer } from '@/__tests__/render';
-import SourceControlPushModal from '@/components/SourceControlPushModal.ee.vue';
+import SourceControlPushModal from './SourceControlPushModal.vue';
 import { createTestingPinia } from '@pinia/testing';
 import { createEventBus } from '@n8n/utils/event-bus';
 import type { SourceControlledFile } from '@n8n/api-types';
-import { useSourceControlStore } from '@/stores/sourceControl.store';
+import { useSourceControlStore } from '../sourceControl.store';
 import { mockedStore } from '@/__tests__/utils';
 import { VIEWS } from '@/constants';
 import { useTelemetry } from '@/composables/useTelemetry';
