+++ packages/frontend/editor-ui/src/features/sourceControl.ee/sourceControl.constants.ts
@@ -1,0 +1,2 @@
+export const SOURCE_CONTROL_PUSH_MODAL_KEY = 'sourceControlPush';
+export const SOURCE_CONTROL_PULL_MODAL_KEY = 'sourceControlPull';
