+++ packages/frontend/editor-ui/src/features/sourceControl.ee/sourceControl.store.test.ts
@@ -1,12 +1,12 @@
 import { createPinia, setActivePinia } from 'pinia';
 import { vi } from 'vitest';
 
-import * as vcApi from '@/api/sourceControl';
-import { useSourceControlStore } from '@/stores/sourceControl.store';
-import type { SourceControlPreferences } from '@/types/sourceControl.types';
+import * as vcApi from './sourceControl.api';
+import { useSourceControlStore } from './sourceControl.store';
+import type { SourceControlPreferences } from './sourceControl.types';
 import type { SourceControlledFile } from '@n8n/api-types';
 
-vi.mock('@/api/sourceControl');
+vi.mock('@/features/sourceControl.ee/sourceControl.api');
 
 vi.mock('@n8n/stores/useRootStore', () => ({
 	useRootStore: vi.fn(() => ({
