+++ packages/frontend/editor-ui/src/features/sourceControl.ee/sourceControl.store.ts
@@ -3,8 +3,8 @@ import { defineStore } from 'pinia';
 import { EnterpriseEditionFeature } from '@/constants';
 import { useSettingsStore } from '@/stores/settings.store';
 import { useRootStore } from '@n8n/stores/useRootStore';
-import * as vcApi from '@/api/sourceControl';
-import type { SourceControlPreferences, SshKeyTypes } from '@/types/sourceControl.types';
+import * as vcApi from './sourceControl.api';
+import type { SourceControlPreferences, SshKeyTypes } from './sourceControl.types';
 import type { TupleToUnion } from '@/utils/typeHelpers';
 import type { SourceControlledFile } from '@n8n/api-types';
