+++ packages/frontend/editor-ui/src/features/sourceControl.ee/views/SettingsSourceControl.test.ts
@@ -4,8 +4,8 @@ import userEvent from '@testing-library/user-event';
 import { createPinia, setActivePinia } from 'pinia';
 import { setupServer } from '@/__tests__/server';
 import { useSettingsStore } from '@/stores/settings.store';
-import { useSourceControlStore } from '@/stores/sourceControl.store';
-import SettingsSourceControl from '@/views/SettingsSourceControl.vue';
+import { useSourceControlStore } from '../sourceControl.store';
+import SettingsSourceControl from './SettingsSourceControl.vue';
 import { createComponentRenderer } from '@/__tests__/render';
 import { EnterpriseEditionFeature } from '@/constants';
 import { nextTick } from 'vue';
