+++ packages/frontend/editor-ui/src/features/sourceControl.ee/views/SettingsSourceControl.vue
@@ -6,8 +6,8 @@ import { useMessage } from '@/composables/useMessage';
 import { usePageRedirectionHelper } from '@/composables/usePageRedirectionHelper';
 import { useToast } from '@/composables/useToast';
 import { MODAL_CONFIRM } from '@/constants';
-import { useSourceControlStore } from '@/stores/sourceControl.store';
-import type { SshKeyTypes, SourceControlPreferences } from '@/types/sourceControl.types';
+import { useSourceControlStore } from '../sourceControl.store';
+import type { SshKeyTypes, SourceControlPreferences } from '../sourceControl.types';
 import type { TupleToUnion } from '@/utils/typeHelpers';
 import type { Rule, RuleGroup } from '@n8n/design-system/types';
 import { useI18n } from '@n8n/i18n';
