+++ packages/frontend/editor-ui/src/stores/projects.store.ts
@@ -18,7 +18,7 @@ import { STORES } from '@n8n/stores';
 import { useUsersStore } from '@/stores/users.store';
 import { getResourcePermissions } from '@n8n/permissions';
 import type { CreateProjectDto, UpdateProjectDto } from '@n8n/api-types';
-import { useSourceControlStore } from '@/stores/sourceControl.store';
+import { useSourceControlStore } from '@/features/sourceControl.ee/sourceControl.store';
 
 export type ResourceCounts = {
 	credentials: number;
