+++ packages/frontend/editor-ui/src/views/CredentialsView.vue
@@ -22,7 +22,7 @@ import { useExternalSecretsStore } from '@/features/externalSecrets/externalSecr
 import { useNodeTypesStore } from '@/stores/nodeTypes.store';
 import { useProjectsStore } from '@/stores/projects.store';
 import { useSettingsStore } from '@/stores/settings.store';
-import { useSourceControlStore } from '@/stores/sourceControl.store';
+import { useSourceControlStore } from '@/features/sourceControl.ee/sourceControl.store';
 import { listenForModalChanges, useUIStore } from '@/stores/ui.store';
 import { useUsersStore } from '@/stores/users.store';
 import type { Project } from '@/types/projects.types';
